Move mapping definition into index classes
